feat: rename identifier to appIdentifier
