+++ packages/aragon-wrapper/src/rpc/handlers/index.js
@@ -47,13 +47,13 @@ export function combineRequestHandlers (...handlers) {
 // Export request handlers
 export { default as accounts } from './accounts'
 export { default as addressIdentity } from './addressIdentity'
+export { default as appIdentifier } from './app-identifier'
 export { default as cache } from './cache'
 export { default as call } from './call'
 export { default as describeScript } from './describe-script'
 export { call as externalCall } from './external'
 export { events as externalEvents } from './external'
 export { default as events } from './events'
-export { default as identifier } from './identifier'
 export { default as intent } from './intent'
 export { default as network } from './network'
 export { default as notifications } from './notifications'
